Refactor id datatype of new stories to from int to long

+++ src/main/java/org/example/topprogramingstoriesdashboard/scraper/NewStoriesClient.java
@@ -15,7 +15,7 @@ public class NewStoriesClient implements NewStoriesGateway {
     }
 
     @Override
-    public List<Integer> getNewStories() {
+    public List<Long> getNewStories() {
         return RestClient
                 .builder()
                 .baseUrl(baseUrl)
